rename and clean up comments
