+++ internal/admission/mutator.go
@@ -1,10 +1,10 @@
-package handlers
+package admission
 
 import (
 	"github.com/kanopy-platform/hedgetrimmer/internal/limitrange"
 	corev1 "k8s.io/api/core/v1"
 )
 
-type PtsMutator interface {
+type PodTemplateSpecMutator interface {
 	Mutate(inputPts corev1.PodTemplateSpec, limitRangeMemory limitrange.Config) (corev1.PodTemplateSpec, error)
 }
