+++ internal/quantity/quantity.go
@@ -75,14 +75,6 @@ func Max(x resource.Quantity, y resource.Quantity) resource.Quantity {
 	return xCopy
 }
 
-// convenience function for testing only
 func Ptr(q resource.Quantity) *resource.Quantity {
 	return &q
-}
-
-// convenience function for testing only
-func RoundUp(q resource.Quantity, scale resource.Scale) resource.Quantity {
-	rounded := q.DeepCopy()
-	rounded.RoundUp(scale)
-	return rounded
 }
